Move query exceptions module
